Preliminary data exploration
Wrangle data to include our classifier and predictors only.
Mutate the data so that final grade (G3) is a factor, and romantic relationship is a numerical value instead of Y/N.
Split the data into a training and testing set.
Explore the testing set, vizualize the distribution of our predictors using histograms